+++ src/Adapter/Route/Object/GetChildren/GetChildrenByRefIdRoute.php
@@ -54,6 +54,21 @@ class GetChildrenByRefIdRoute implements Route
             ],
             [
                 RouteParamDocumentationDto::new(
+                    "id",
+                    "int",
+                    "Filter object to id"
+                ),
+                RouteParamDocumentationDto::new(
+                    "import_id",
+                    "string",
+                    "Filter object to import id"
+                ),
+                RouteParamDocumentationDto::new(
+                    "ref_id",
+                    "int",
+                    "Filter object to ref id"
+                ),
+                RouteParamDocumentationDto::new(
                     "types",
                     ObjectType::class . "[]",
                     "Filter by object types split by ,"
@@ -104,6 +119,15 @@ class GetChildrenByRefIdRoute implements Route
     {
         $children = $this->ilias_rest_api->getChildrenByRefId(
             $request->getParam(
+                "ref_id"
+            ),
+            $request->getQueryParam(
+                "id"
+            ),
+            $request->getQueryParam(
+                "import_id"
+            ),
+            $request->getQueryParam(
                 "ref_id"
             ),
             ($types = $request->getQueryParam(