+++ src/Service/Object/Port/ObjectService.php
@@ -291,7 +291,7 @@ class ObjectService
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenById(int $id, ?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenById(int $id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return GetChildrenCommand::new(
             $this,
@@ -299,8 +299,11 @@ class ObjectService
         )
             ->getChildrenById(
                 $id,
-                $types,
-                $title,
+                $children_id,
+                $children_import_id,
+                $children_ref_id,
+                $children_types,
+                $children_title,
                 $ref_ids,
                 $in_trash
             );
@@ -311,7 +314,7 @@ class ObjectService
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenByImportId(string $import_id, ?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenByImportId(string $import_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return GetChildrenCommand::new(
             $this,
@@ -319,8 +322,11 @@ class ObjectService
         )
             ->getChildrenByImportId(
                 $import_id,
-                $types,
-                $title,
+                $children_id,
+                $children_import_id,
+                $children_ref_id,
+                $children_types,
+                $children_title,
                 $ref_ids,
                 $in_trash
             );
@@ -331,7 +337,7 @@ class ObjectService
      * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenByRefId(int $ref_id, ?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenByRefId(int $ref_id, ?int $children_id = null, ?string $children_import_id = null, ?int $children_ref_id = null, ?array $children_types = null, ?string $children_title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return GetChildrenCommand::new(
             $this,
@@ -339,8 +345,11 @@ class ObjectService
         )
             ->getChildrenByRefId(
                 $ref_id,
-                $types,
-                $title,
+                $children_id,
+                $children_import_id,
+                $children_ref_id,
+                $children_types,
+                $children_title,
                 $ref_ids,
                 $in_trash
             );
@@ -387,12 +396,15 @@ class ObjectService
      * @param ObjectType[]|null $types
      * @return ObjectDto[]
      */
-    public function getObjects(?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : array
+    public function getObjects(?int $id = null, ?string $import_id = null, int $ref_id = null, ?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : array
     {
         return GetObjectsCommand::new(
             $this->ilias_database
         )
             ->getObjects(
+                $id,
+                $import_id,
+                $ref_id,
                 $types,
                 $title,
                 $ref_ids,