+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-05-08-1
+
+Changes:
+
+- Object ids filter
+
 ## v2023-05-05-1
 
 Changes: